Fixed conversion from 'double' to 'int' warning

+++ Breeder/BR_Misc.cpp
@@ -557,8 +557,8 @@ void MidiItemTempo (COMMAND_T* ct)
 			else
 			{
 				BR_MidiItemTimePos timePos(item, false);
-				int timeBase = GetMediaItemInfo_Value(item, "C_BEATATTACHMODE");
-				SetMediaItemInfo_Value(item, "C_BEATATTACHMODE", 0);
+				double timeBase = GetMediaItemInfo_Value(item, "C_BEATATTACHMODE");
+				SetMediaItemInfo_Value(item, "C_BEATATTACHMODE", 0.0);
 				if (SetIgnoreTempo(item, ignoreTempo, bpm, num, den, true))
 				{
 					timePos.Restore(true);
